Added error messages as embedded list in page. Decided on this after looking over several Vue libs for snackbars and alert solutions. Most run into Material Design recommendation to not show more than one at a time, which is just not terribly useful.
It's summed up well here: https://github.com/vuetifyjs/vuetify/issues/2384

In the end, went with a simple, global list of alerts and a utility to help show/hide them.

It's a bit problematic because it moves the content on the screen. Might look at making them an overlay of some sort later to avoid misclicks from moving content.