Move widgets into components and break down into subdirectories. - Naming schema is one word no camel case; considered _ and discarded. - @/widgets is now part of @/components - Added @/components/room (components specific to Room.vue) - Added @/components/room/eventcard (all the cards shown in the feed) - Added @/components/uiwidget (pure ui components with no logic)
Justification: I couldn't answer the question, "What is the difference
between a component and a widget?" There's really no logical difference;
it was just a way to try and break up the contents so it's not a giant
list. At some point it occurred to me that subdirectories and logical
topics was a better organization strategy. Welcome to Kato's brain.