Room toolbars can now collapse into the dropdown pancake. - Renamed AdminDropdown to RoomToolbarDropdown (and it now works for admins and other users) - Abstracted reaction emojis into their own component - Fixed a bug in Debug.vue caused by recursive json due to sharedScope having Auth objects in it.
This allows additional emojis and actions to be shown on wider views and to be in the dropdown on smaller ones. It also fixes a bug where the menu would expand beyond the right side of its container.

Also made Room layout fluid which made things far nicer overall during resize.