Refactored poll to be under feed/{event_id}/meta/poll instead of polls/poll.
Why this choice? There can only be one poll per event. The poll doesn't need its own collection just for the sake of storing its document. We could have added it directly on the event as an object, but that is hard to monitor and pass around in a modular manner. Instead, putting it in its own doc, but under the meta/ path allows it to live next to meta/wait and any other event types we decide to build metadata for.